requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
firebase-admin==6.5.0
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.0
